Header file cleanup
