+++ .gitignore
@@ -1,3 +1,4 @@
 libantic*
 Makefile
 build/
+antic
